move message code to message.go
